requirements.txt
python-dotenv==1.0.1
requests==2.32.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.133.0
tzdata==2024.1
PyNaCl==1.5.0
